requirements.txt
Flask==3.1.2
SQLAlchemy==2.0.44
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
psycopg2-binary==2.9.11
flask-cors==6.0.1
marshmallow==4.1.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
python-dotenv==1.2.1
requests==2.32.5
uWSGI==2.0.31